chore: move back getTypeIdentifier to common/types to avoid exposing it to the public API
