+++ src/hooks/useModule.ts
@@ -3,15 +3,7 @@ import { fetchResource } from '../utils/fetchResource';
 import { ETError, getError } from '../Error';
 import { ETInput, Module } from '../types/common';
 import { _ETModule } from '../native/RnExecutorchModules';
-
-export const getTypeIdentifier = (input: ETInput): number => {
-  if (input instanceof Int8Array) return 1;
-  if (input instanceof Int32Array) return 3;
-  if (input instanceof BigInt64Array) return 4;
-  if (input instanceof Float32Array) return 6;
-  if (input instanceof Float64Array) return 7;
-  return -1;
-};
+import { getTypeIdentifier } from '../types/common';
 
 interface Props {
   modelSource: string | number;
